Finished and test report written
I have opted out of adding the number of comments user each slide as this would either involve a fresh server read at each step and load to excess server reads, it would also require a new variable within the main document that would need to be updated with each extra comment and could lead to crashes where multiple users and adding and removing comments. Alternatively, it would need a full server read for the number of documents in the collection which could lead to unnecessary exponential server reads.

For this reason, users are required to open the comments section to see the comments available.