Fixed timer bugs, added new UI tests and added new helpful navigation to recipe helper
Fixed bug in timer where it would display on the first slide.
Added new UI tests for continuity, and fixed issue where timer was counting down instead of up and for values greater than 60 seconds was displaying in-correctly.

Also added a super helpful navigation thingy into EspressoHelper. Gif showing this off is in the URL below:

https://imgur.com/FSkUbNr