Design changes
Lots of design changes, mainly colours.
Happy with everything except the profile page and I can't find where to change the background colour for reviews.

There is a very fine line between the app looking like the inside of a hospital and an app that people would enjoy looking at.
Took inspiration from the trainline app and the simplyvegan app on IOS.
My phone is pretty old and cheap so the colours compared to the ipad are quite muted so a lot of it's a best assumption.